% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{baseName}
\alias{baseName}
\title{Create template name}
\usage{
baseName(author, year)
}
\arguments{
\item{author}{a string representing the author's first and last name}

\item{year}{a string representing the publication year}
}
\value{
PMID or ""
}
\description{
Create template name
}
\details{
The name is the author's last name plus the first initial
}
\examples{
baseName("Grace Brewster Hopper", "2025")
}
