% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioproject_to_biosample_metadata.R
\name{bioproject_to_biosample_metadata}
\alias{bioproject_to_biosample_metadata}
\title{BioProject ID to BioSample Metadata Converter}
\usage{
bioproject_to_biosample_metadata(bioproject_id, max_samples = NULL)
}
\arguments{
\item{bioproject_id}{}

\item{max_samples}{}
}
\description{
This script retrieves BioSample metadata from BioProject IDs using NCBI E-utilities
}
\examples{
# Single BioProject
metadata <- bioproject_to_biosample_metadata("PRJNA13")

# Limit to first 50 samples
metadata <- bioproject_to_biosample_metadata("PRJEB9576", max_samples = 50)

# Multiple BioProjects
bioproject_ids <- c("PRJNA13", "PRJEB9576")
all_metadata <- lapply(bioproject_ids, function(id) {
  bioproject_to_biosample_metadata(id, max_samples = 50)
})
combined_metadata <- bind_rows(all_metadata)

}
