% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkCuration.R
\name{checkCuration}
\alias{checkCuration}
\title{Title Check the curation of per-participant metadata against a template.}
\usage{
checkCuration(curated, template = read.csv(system.file("extdata/template.csv",
  package = "curatedMetagenomicDataCuration"), as.is = TRUE))
}
\arguments{
\item{curated}{a data.frame containing the curated per-participant metadata to be checked}

\item{template}{a data.frame defining valid syntax}
}
\value{
a list with two elements, [["colnames"]] and [["values]]
}
\description{
Title Check the curation of per-participant metadata against a template.
}
\details{
See Examples for the template used for curatedMetagenomicData. 
The template has five columns:

1. "col.name" specifies the name of the column in the curated data.frame.
2. "multiplevalues": multiple semicolon-separated values are allowed
3. "uniqueness": unique means each value must be unique, non-unique means repeated values are allowed
4. "requiredness": if "required", there must be no missing (NA) values. If "optional", missing values are allowed.
5. "allowedvalues": a regex defining allowable values for the column
6. "description": a free-form description of the variable
}
