% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{encodeAndClean}
\alias{encodeAndClean}
\title{Replace quotes with HTML code, remove HTML and line breaks}
\usage{
encodeAndClean(s)
}
\arguments{
\item{s}{a string possibly with HTML, quotes, and line breaks}
}
\value{
string with HTML and line breaks removed, quotes encoded
}
\description{
Replace quotes with HTML code, remove HTML and line breaks
}
\details{
See https://stackoverflow.com/questions/17227294/removing-html-tags-from-a-string-in-r#answer-17227415
}
\examples{
encodeAndClean("There's two lines.\nThere's also <b>HTML</b>.")
}
