% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_validation_helpers.R
\name{find_metadata_files}
\alias{find_metadata_files}
\title{Find Metadata Files}
\usage{
find_metadata_files(path = "inst/harmonized/")
}
\arguments{
\item{path}{Base path to search (default: "inst/harmonized/")}
}
\value{
Character vector of file paths
}
\description{
Find all *_sample.tsv files in harmonized directory
}
