% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_report_generators.R
\name{generate_console_report}
\alias{generate_console_report}
\title{Generate Console Report}
\usage{
generate_console_report(validation_summary)
}
\arguments{
\item{validation_summary}{Aggregated validation results}
}
\description{
Print validation results to console
}
