% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_metadata.R
\name{get_metadata}
\alias{get_metadata}
\title{Get relevant metadata for each SRR using SRP}
\usage{
get_metadata(srp)
}
\arguments{
\item{srp}{An SRP as a string beginning with the prefix "SRP"}
}
\value{
a data.frame with two elements, [["colnames"]] and [["values]]
}
\description{
Get relevant metadata for each SRR using SRP
}
\details{
See Examples for the template used for curatedMetagenomicData. 
The template has five columns:

1. "col.name" specifies the name of the column in the curated data.frame.
2. "multiplevalues": multiple semicolon-separated values are allowed
3. "uniqueness": unique means each value must be unique, non-unique means repeated values are allowed
4. "requiredness": if "required", there must be no missing (NA) values. If "optional", missing values are allowed.
5. "allowedvalues": a regex defining allowable values for the column
6. "description": a free-form description of the variable
}
