% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getCrossref}
\alias{getCrossref}
\title{Get data from Crossref}
\usage{
getCrossref(doi)
}
\arguments{
\item{doi}{A string representing a DOI}
}
\value{
list with data associated with the DOI
}
\description{
Get data from Crossref
}
\examples{
getCrossref("10.1128/mSystems.00164-16")
}
