% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getPattern}
\alias{getPattern}
\title{Get pattern}
\usage{
getPattern(ss, patternName)
}
\arguments{
\item{ss}{a string}

\item{patternName}{a string matching a name in .PATTERN}
}
\value{
a string matching the pattern
}
\description{
Get pattern
}
\examples{
ss <- "The DOI: 10.1128/mSystems.00164-16\nBioProject ID: PRJEB1234"
getPattern(ss, "doi")
getPattern(ss, "bioproject")
}
