% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getPmid}
\alias{getPmid}
\title{Get PMID from Eutils}
\usage{
getPmid(doi)
}
\arguments{
\item{doi}{a string representing a DOI}
}
\value{
PMID or ""
}
\description{
Get PMID from Eutils
}
\examples{
getPmid("10.1128/mSystems.00164-16")
}
