% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_long_form.R
\name{metadata_long_form}
\alias{metadata_long_form}
\title{convert all metadata.tsv files in package to long form}
\usage{
metadata_long_form()
}
\value{
a data.frame with the following columns:

- study
- sampleID
- subjectID
- variable: the column name from the wide-form data
- numeric_value: if a numeric value, this value will be non-NA
- string_value: if a character value, this value will be non-NA
- logical_value: if a logical value, this value will be non-NA
}
\description{
This function runs through the metadata.tsv files in 
the installed package, running `study_long_form()` on
each, then binds them together, appending a `study` column.
}
\examples{
ret = metadata_long_form()
dim(ret)
ret
table(ret$study)
summary(ret)

}
\author{
Sean Davis <seandavi@gmail.com>
}
