% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_statistics_collectors.R
\name{normalize_values}
\alias{normalize_values}
\title{Normalize Values}
\usage{
normalize_values(x)
}
\arguments{
\item{x}{Character vector}
}
\value{
Cleaned character vector
}
\description{
Clean and normalize metadata values
}
