% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odm_get_samples.R
\name{odm_get_samples}
\alias{odm_get_samples}
\title{Get samples by Genestack study accession}
\usage{
odm_get_samples(
  host = "https://odm.drylab.tech",
  token,
  study_id = NULL,
  page_limit = 100,
  offset = 0
)
}
\arguments{
\item{host}{The host URL for the Genestack API}

\item{token}{The API token for authentication}

\item{study_id}{A character(1). Genestack accession id for the study.}

\item{page_limit}{The number of samples to retrieve per page 
(default: 100). The maximum is 2,000.}

\item{offset}{Show the page {pageOffset+1} results from the start of 
the results. (e.g., 100 will show a page of results starting from the 
101st result (default: 0). To return all results iterate through pages 
using `offset` values of `n * page_limit` until it surpaces the total
number of samples.}
}
\value{
A data frame containing sample metadata. A message shows how many 
samples are retrieved out of the total samples.
}
\description{
This function retrieves samples from the Genestack API, filters them by 
Genestack study accession.
}
