% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odm_get_studies.R
\name{odm_get_studies}
\alias{odm_get_studies}
\title{Get the list of studies in ODM}
\usage{
odm_get_studies(api_token)
}
\arguments{
\item{api_token}{ODM API token}
}
\value{
A data frame containing study-level metadata for all the studies 
a user as access to in ODM.
}
\description{
Get the list of studies in ODM
}
