% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PubMed.R
\name{PubMed}
\alias{PubMed}
\title{Title Retrieve article details from PubMed}
\usage{
PubMed(PMID)
}
\arguments{
\item{pmid}{A valid PMID as a string}
}
\value{
a PubMed-class object
}
\description{
Title Retrieve article details from PubMed
}
\details{
Retrieves article details for a PMID and returns a PubMed class object.
The object is a list with three data.frames:

1. pubinfo contains details of the article and its publication
2. authors contains the author list
3. affiliations contains the unique author affiliations
}
