% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_statistics_collectors.R
\name{split_values}
\alias{split_values}
\title{Split Multi-Value Fields}
\usage{
split_values(x)
}
\arguments{
\item{x}{Character vector}
}
\value{
Character vector with split values
}
\description{
Split fields with multiple values (semicolon/pipe/comma separated)
}
