% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_long_form.R
\name{study_long_form}
\alias{study_long_form}
\title{read a metadata tsv file and convert to long form}
\usage{
study_long_form(fname)
}
\arguments{
\item{fname}{character(1) giving a path to one of the 
metadata.tsv files.}
}
\value{
a 6-column dataframe
}
\description{
This is basically an auxilliary function for converting
datasets from the wide-format metadata to long-form, or "tidy"
format.
}
\seealso{
metadata_long_form()
}
\author{
Sean Davis <seandavi@gmail.com>
}
