% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_statistics_collectors.R
\name{update_counts}
\alias{update_counts}
\title{Update Value Counts}
\usage{
update_counts(counts, values)
}
\arguments{
\item{counts}{Named integer vector}

\item{values}{Character vector of values}
}
\value{
Updated counts
}
\description{
Update count table with new values
}
