% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{createTemplate}
\alias{createTemplate}
\title{Create template}
\usage{
createTemplate(base, path, overwrite = FALSE)
}
\arguments{
\item{path}{a string}

\item{author}{a string representing the author's first and last name}

\item{year}{a string representing the publication year}
}
\value{
invisible NULL
}
\description{
Create template
}
\details{
Note: the provided path must already exist otherwise the function
will fail.

The name is the author's last name plus the first initial
}
\examples{
createTemplate("HopperG_2025", tempdir())
}
