% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.getRequest}
\alias{.getRequest}
\title{Get request}
\usage{
.getRequest(url)
}
\arguments{
\item{url}{a string representing a URL}
}
\value{
a string representing JSON response
}
\description{
Get request
}
\examples{
getRequest(.CROSSREF("10.1128/mSystems.00164-16"))
}
