% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_metadata.R
\name{get_metadata}
\alias{get_metadata}
\title{Get relevant metadata for each SRR using SRP}
\usage{
get_metadata(srp)
}
\arguments{
\item{srp}{An SRP as a string beginning with the prefix "SRP"}
}
\value{
a data.frame with sampleID, sequencing_platform, number_reads, and avg_read_length
}
\description{
Get relevant metadata for each SRR using SRP
}
\details{
Supports studies with up to 99,990 runs. See Examples for the template used for curatedMetagenomicData. 
The data.frame has five columns:

1. "sampleID" Sample ID.
2. "sequencing_platform": Sequencing platform used.
3. "number_reads": Number of reads. If there are multiple SRRs, the sum of number_reads is used.
4. "avg_read_length": Average read length If there are multiple SRRs, the mean avg_read_length  is used.
5. "SRRs": A list of SRRs associated with the sample ID.
}
