% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_validation_helpers.R
\name{load_validation_schema}
\alias{load_validation_schema}
\title{Load Validation Schema}
\usage{
load_validation_schema()
}
\value{
List containing the loaded schema
}
\description{
Load the metadata schema from OmicsMLRepoCuration package
}
