% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeCombinedMetadata.R
\name{makeCombinedMetadata}
\alias{makeCombinedMetadata}
\title{Make a combined metadata table from individual TSVs}
\usage{
makeCombinedMetadata()
}
\value{
A data.table containing metadata for all studies in inst/curated
}
\description{
Make a combined metadata table from individual TSVs
}
