% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nf_get_completed_run.R
\name{nf_get_completed_run}
\alias{nf_get_completed_run}
\title{Get Completed Runs from cMD MetaPhlAn4 Nextflow Telemetry API}
\usage{
nf_get_completed_run(limit = 250)
}
\arguments{
\item{limit}{Integer. Maximum number of runs to retrieve. Default is 250.}
}
\value{
A data frame containing the API response. There are four columns
on 'run_id', 'run_name', 'utc_time', and 'sample_id'.
}
\description{
This function makes an API request to the Nextflow telemetry service
to retrieve information about completed runs.
}
\examples{
# Get only 10 completed runs
runs <- nf_get_completed_run(limit = 10)

}
